work in progress
Most of the database functions are complete, though I need to clean up and remove some redundant methods. Currently working on presentation by building in the pandas dataframe. This will help with navigating, adding, and removing from the database. Once these are more established, I'll write a main file to run the program.